+++ Client/Calculator/calculator.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Calculator</title>
-    <link rel="stylesheet" href="/css/calculator.css">
+    <link rel="stylesheet" href="./calculator.css">
     <link
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.3.0/css/font-awesome.css"
       rel="stylesheet"
@@ -14,7 +14,6 @@
       href="https://fonts.googleapis.com/css?family=Inter"
       rel="stylesheet"
     />
-    <script src="js/calculator.js" defer></script>
 </head>
 <body>
     <main id="calculator-content">
@@ -43,5 +42,6 @@
             <button class="calculator-button" type="button" value="=">=</button>
         </form>
     </main>
+    <script src="calculator.js"></script>
 </body>
 </html>
