+++ Client/Trivia/index.html
@@ -17,9 +17,8 @@
       href="https://fonts.googleapis.com/css?family=Inter"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="css/trivia.css" />
-    <script type="text/javascript" src="js/trivia.js" defer></script>
-    <script type="text/javascript" src="js/logout.js" defer></script>
+    <link rel="stylesheet" href="./trivia.css" />
+    <script type="text/javascript" src="./trivia.js" defer></script>
   </head>
   <body>
     <main id="home-content">
@@ -36,7 +35,6 @@
         <button class="navigation-button" id="try-again">Try Again</button>
         <button class="navigation-button">Calculator</button>
       </section>
-      <button id="logout-button">Logout</button>
     </main>
   </body>
 </html>
