+++ calculator_quiz/src/main/resources/data.sql
