--- client/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "Math Trivia",
-    "version": "1.0.0",
-    "description": "Math trivia",
-    "main": "app.js",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "node ./app.js"
-    },
-    "dependencies": {
-        "express": "^4.18.2",
-        "cors": "^2.8.5"
-    }
-}
